Changes w.r.t requirement

+++ Weather/Weather/Model/WeatherReport.cs
@@ -13,7 +13,7 @@ namespace Weather.Model
 
     public class Weather
     {
-        public int id { get; set; }
+        public double id { get; set; }
         public string main { get; set; }
         public string description { get; set; }
         public string icon { get; set; }
@@ -22,8 +22,8 @@ namespace Weather.Model
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
@@ -31,23 +31,23 @@ namespace Weather.Model
     public class Wind
     {
         public double speed { get; set; }
-        public int deg { get; set; }
+        public double deg { get; set; }
         public double gust { get; set; }
     }
 
     public class Clouds
     {
-        public int all { get; set; }
+        public double all { get; set; }
     }
 
     public class Sys
     {
-        public int type { get; set; }
-        public int id { get; set; }
+        public double type { get; set; }
+        public double id { get; set; }
         public double message { get; set; }
         public string country { get; set; }
-        public int sunrise { get; set; }
-        public int sunset { get; set; }
+        public double sunrise { get; set; }
+        public double sunset { get; set; }
     }
 
     public class WeatherReport
@@ -56,13 +56,13 @@ namespace Weather.Model
         public List<Weather> weather { get; set; }
         public string @base { get; set; }
         public Main main { get; set; }
-        public int visibility { get; set; }
+        public double visibility { get; set; }
         public Wind wind { get; set; }
         public Clouds clouds { get; set; }
-        public int dt { get; set; }
+        public double dt { get; set; }
         public Sys sys { get; set; }
-        public int id { get; set; }
+        public double id { get; set; }
         public string name { get; set; }
-        public int cod { get; set; }
+        public double cod { get; set; }
     }
 }
